auth_service/requirements.txt
Django==5.1.7
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
dotenv